Fix optimizer and optimzier hook tests
