+++ tests/chainer_tests/test_optimizer.py
@@ -6,7 +6,6 @@ import numpy as np
 import chainer
 from chainer import cuda
 from chainer import optimizer
-from chainer import optimizers
 from chainer import testing
 from chainer.testing import attr
 
@@ -95,137 +94,6 @@ class SimpleLink(chainer.Link):
         super(SimpleLink, self).__init__(param=w.shape)
         self.param.data = w
         self.param.grad = g
-
-
-class TestOptimizerWeightDecay(unittest.TestCase):
-
-    def setUp(self):
-        self.target = SimpleLink(
-            np.arange(6, dtype=np.float32).reshape(2, 3),
-            np.arange(3, -3, -1, dtype=np.float32).reshape(2, 3))
-
-    def check_weight_decay(self):
-        w = self.target.param.data
-        g = self.target.param.grad
-
-        decay = 0.2
-        expect = w - g - decay * w
-
-        opt = optimizers.SGD(lr=1)
-        opt.setup(self.target)
-        opt.add_hook(optimizer.WeightDecay(decay))
-        opt.update()
-
-        testing.assert_allclose(expect, w)
-
-    def test_weight_decay_cpu(self):
-        self.check_weight_decay()
-
-    @attr.gpu
-    def test_weight_decay_gpu(self):
-        self.target.to_gpu()
-        self.check_weight_decay()
-
-
-class TestOptimizerLasso(unittest.TestCase):
-
-    def setUp(self):
-        self.target = SimpleLink(
-            np.arange(6, dtype=np.float32).reshape(2, 3),
-            np.arange(3, -3, -1, dtype=np.float32).reshape(2, 3))
-
-    def check_lasso(self):
-        w = self.target.param.data
-        g = self.target.param.grad
-        xp = cuda.get_array_module(w)
-        decay = 0.2
-        expect = w - g - decay * xp.sign(w)
-
-        opt = optimizers.SGD(lr=1)
-        opt.setup(self.target)
-        opt.add_hook(optimizer.Lasso(decay))
-        opt.update()
-
-        testing.assert_allclose(expect, w)
-
-    def test_lasso_cpu(self):
-        self.check_lasso()
-
-    @attr.gpu
-    def test_lasso_gpu(self):
-        self.target.to_gpu()
-        self.check_lasso()
-
-
-class TestOptimizerGradientNoise(unittest.TestCase):
-
-    eta = 0.01
-
-    def setUp(self):
-        self.target = SimpleLink(
-            np.arange(6, dtype=np.float32).reshape(2, 3),
-            np.arange(3, -3, -1, dtype=np.float32).reshape(2, 3))
-
-        self.noise_value = np.random.normal(
-            loc=0, scale=np.sqrt(self.eta / np.power(1, 0.55)),
-            size=(2, 3)).astype(np.float32)
-
-    def check_gradient_noise(self):
-        w = self.target.param.data
-        g = self.target.param.grad
-        xp = cuda.get_array_module(w)
-        noise_value = xp.asarray(self.noise_value)
-
-        expect = w - g - noise_value
-
-        noise = mock.Mock(return_value=noise_value)
-        opt = optimizers.SGD(lr=1)
-        opt.setup(self.target)
-        hook = optimizer.GradientNoise(self.eta, noise_func=noise)
-        opt.add_hook(hook)
-        opt.update()
-
-        testing.assert_allclose(expect, w, rtol=0.4)
-        noise.assert_called_once_with(xp, (2, 3), np.float32, hook, opt)
-
-    def test_gradient_noise_cpu(self):
-        self.check_gradient_noise()
-
-    @attr.gpu
-    def test_gradient_noise_gpu(self):
-        self.target.to_gpu()
-        self.check_gradient_noise()
-
-
-class TestGradientHardClipping(unittest.TestCase):
-
-    def setUp(self):
-        self.target = SimpleLink(
-            np.arange(6, dtype=np.float32).reshape(2, 3),
-            np.arange(3, -3, -1, dtype=np.float32).reshape(2, 3))
-
-    def check_hardclipping(self):
-        w = self.target.param.data
-        g = self.target.param.grad
-        xp = cuda.get_array_module(w)
-        lower_bound = -0.9
-        upper_bound = 1.1
-        expect = w - xp.clip(g, lower_bound, upper_bound)
-
-        opt = optimizers.SGD(lr=1)
-        opt.setup(self.target)
-        opt.add_hook(optimizer.GradientHardClipping(lower_bound, upper_bound))
-        opt.update()
-
-        testing.assert_allclose(expect, w)
-
-    def test_hardclipping_cpu(self):
-        self.check_hardclipping()
-
-    @attr.gpu
-    def test_hardclipping_gpu(self):
-        self.target.to_gpu()
-        self.check_hardclipping()
 
 
 class TestGradientMethod(unittest.TestCase):
